requirements.txt
python-whois==0.7.3
requests==2.31.0
python-dateutil==2.8.2
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==0.21.0
